Adapt Packer build for K3s and improve vault workflow
We can have Packer just prompt for the vault password as-an-when it
needs it, however the downside of this is the necessity of watching the
build as it happens: the Ansible provisioning step takes place after
downloading and mounting the ISO in a chroot, and hence the vault
password is only prompted for ~5 minutes after starting the build. To
make the build 'hands-free', i.e. such that we can walk away from it,
the build script now prompts directly.

Note that we must messily use a temp file for this (though the user can
set the `ANSIBLE_VAULT_PASSWORD_FILE` to avoid this!), because we can't
(and indeed shouldn't) pass the Vault password directly on the
command-line.