Install ArgoCD and create manifests skeleton
To keep with the hands-off-setup approach, ArgoCD is now automagically
installed by utilising K3s addons. In addition, we set up the root
Application with which we implement the App-of-Apps pattern.

The current version of ArgoCD has a neat feature: ApplicationSets, which
mean that we can avoid duplicating directory names and having a big long
list of apps, and instead we can just have it load all directories as
separate Applications! Hence, our root App simply contains an
ApplicationSet.