Work around VSO cross-namespace issue
See https://github.com/hashicorp/vault-secrets-operator/issues/423

The VSO looks for the OAuth secret in the same namespace as the
HCPVaultSecretsApp, and there is no way to override this (afaict).

The intended use of the VSO seems to be creating the OAuth secret in
every namespace. To me, this seems bad: I don't want random workloads in
some namespace to have access to my entire Vault.

This workaround attempts to use Reflector to reflect the secret
automatically into the target namespace. The HCPVaultSecretsApp is then
created in the vault-secrets-operator namespace.