Add cloud-init for first-boot configuration tasks
Thus-far we have hacked around first-boot configuration with systemd
units and the like. Cloud-init has modules to do many of the tasks we
want to do.

In future, we could consider replacing the systemd units with cloud-init
fully. This would simplify the Ansible side of things somewhat and
reduce the amount of boilerplate.