Add node name suffix for DNS
The node name defaults to the hostname of the node in K3s, and this is
used as the source of the external-dns name when `--fqdn-template` is
not used.

Debian generally discourages using FQDNs as hostname, and hence we
prefer not to set the hostname to an FQDN here. Instead, we can simply
append a suffix to the K3s node name, which feels like a slightly
cleaner solution.