Simplify pwm_fan module
I think I was quite tired when writing this originally: there is no need
for this to be a systemd unit, as the ARM Packer builder does in fact
mount the FAT partition that we'd need to write to as a volume -- we
just had the wrong mountpoint set.

Thus, we can remove all of the systemd cruft, and simply use the
ini_file module to set the requisite line in config.txt.